Update README.md
An elf and her friends defeat a demon king in a great war. But the war is over, and the elf must search for a new way of life. It's humorous, light hearted, thoughtful, philosophical and at the same time sad. Being the youngest and the longest to live in a legendary adventurer group, Freiren would be the only survivor left. Her journey joined by her pupil and her friend's apprentice.